requirements.txt
python_dotenv
azure-ai-translation-text==1.0.0b1